quick fixes on menu page
Changed it to take the identifier for an hour as a parameter for
simplicity. We can also take out hall_id later if we want to, since
it's included in the hour info. This change simplifies the logic for
figuring out exactly which meal that you want and allows for viewing of
future meals instead of being dependent on the current time to view the
menu.